feat(rustlings): learned hashmaps.

+++ rustlings/exercises/08_enums/enums2.rs
@@ -7,10 +7,10 @@ struct Point {
 #[derive(Debug)]
 enum Message {
     // TODO: Define the different variants used below.
-    Resize { width: i32, height: i32 },
+    Resize { width: u64, height: u64 },
     Move(Point),
     Echo(String),
-    ChangeColor(i32, i32, i32),
+    ChangeColor(u8, u8, u8),
     Quit,
 }
